requirements.txt
streamlit
pandas
sqlalchemy
mysql-connector-python
psycopg2-binary